pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pulse"
version = "0.1.0"
description = "Weekly Product Review Pulse — automated App Store & Play Store review insights via MCP"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    # CLI
    "typer>=0.12.0",
    # Config & validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    # HTTP
    "httpx>=0.27.0",
    "requests>=2.32.0",
    # Ingestion
    "google-play-scraper>=1.2.7",
    "langdetect>=1.0.9",
    # Templating
    "jinja2>=3.1.4",
    # MCP
    "mcp>=1.0.0",
    # Utilities
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
reasoning = [
    # Embeddings (local CPU, no API key)
    "sentence-transformers>=3.0.0",
    # Dim reduction + clustering
    "umap-learn>=0.5.6",
    "hdbscan>=0.8.38",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    # Token budget counting
    "tiktoken>=0.7.0",
    "tokenizers>=0.19.0",
    # LLM providers (all free-tier; optional — REST clients are used by default)
    "google-generativeai>=0.8.0",
    "groq>=0.9.0",
]
pii-strict = [
    # Optional deep PII scrubbing (requires Python 3.11-3.13 for spacy wheels).
    "presidio-analyzer>=2.2.354",
    "presidio-anonymizer>=2.2.354",
    "spacy>=3.7.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "types-pyyaml>=6.0.12",
    "types-requests>=2.32.0",
    "types-python-dateutil>=2.9.0",
]
all = [
    "pulse[reasoning]",
    "pulse[pii-strict]",
    "pulse[dev]",
]

[project.scripts]
pulse = "pulse.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/pulse"]

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "config/",
    "tests/",
    "README.md",
    "pyproject.toml",
]

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4", "PIE", "T20", "SIM", "ANN"]
ignore = [
    "ANN401",  # any type allowed in a few spots
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S"]

# ── Mypy ─────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
ignore_missing_imports = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ── Pytest ───────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
pythonpath = ["src"]

# ── Coverage ─────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["pulse"]
branch = true

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:", "raise NotImplementedError"]
